Use an actual domain with UAA LoadBalancer
In its current form, this is problematic because:

A. ClusterIP doesn't work for uaa-public since it won't be publicly addressable
https://kubernetes.io/docs/concepts/services-networking/service/#publishing-services-service-types
We tried using NodePort, but this would result in a randomly generated
30000-32999 port that was addressable via the node's public IP.
Terrible, I know. Using a fixed port wasn't any better, we really
shouldn't be going via K8S node IPs.
We eventually settled on LoadBalancer which works, but only if we
hard-code it in uaa-public, otherwise a lot of other stuff breaks

B. The load balancer gets created dynamically, meaning that the IP is
not known ahead of time. This requires us to configure the DNS after the
service gets created, which is the opossite of what happens in a typical
BOSH scenario: pave the IaaS then create some deployments. We have a
good example of this conventional approach in BOSH Bootloader:
https://github.com/cloudfoundry/bosh-bootloader/blob/master/docs/index.md

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>