Always sync K8S credentials before running kubectl
Otherwise, if the cluster details change, maybe from a different host,
all kubectl commands will timeout after 60s since the IPs and
credentials will no longer match.

This adds a 3s penalty to all commands using kubectl, but the
credentials will always be fresh. This could be done better, but this
solution is good-enough for now.